requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black @ git+https://github.com/psf/black@98a580bbdc217a8b1a47b772d2140ed29cdc587a
dotenv==0.9.9
Faker==37.1.0
fastapi==0.115.12
flake8==7.2.0
idna==3.10
kafka-python==2.1.5
mccabe==0.7.0
py4j==0.10.9.7
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pyspark==3.5.5
python-dotenv==1.1.0
redis==5.2.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
